requirements.txt
cython
numpy==1.23
scipy
tqdm
einops
shapely
pyarrow
timm==0.9.2
h5py
submitit
scikit-image
progressbar2
omegaconf
motmetrics
fairscale
transformers==4.31.0
sentencepiece
accelerate==0.21.0
bitsandbytes
deepspeed==0.10.1
peft==0.5.0


